# use === for comparing characters