one of the items from #282 

A follow up PR might like to do for `Int`s also.