@omus 2020-09-04
Looks good. Would be good to post before/after benchmarks for the record