@omus 2020-09-04
Micro benchmark isn't effective here:
```julia
julia> c = '0'
'0': ASCII/Unicode U+0030 (category Nd: Number, decimal digit)

julia> @btime $c === 'Z'
  0.033 ns (0 allocations: 0 bytes)
false

julia> @btime $c == 'Z'
  0.033 ns (0 allocations: 0 bytes)
false
```